# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CRM Proto API",
    "version": "1.0.0",
    "description": "# CRM Proto API\n\nREST API для CRM системы.\n\n## Аутентификация\n\nAPI использует cookie-based аутентификацию через NextAuth.js.\nДля доступа к защищённым эндпоинтам необходимо авторизоваться через `/login`.\n\n## Ошибки\n\nВсе ошибки возвращаются в формате:\n\n```json\n{\n  \"error\": \"Описание ошибки\"\n}\n```\n\nСтандартные коды ответов:\n- `200` - Успешный запрос\n- `201` - Ресурс создан\n- `400` - Неверный запрос (ошибка валидации)\n- `401` - Не авторизован\n- `403` - Доступ запрещён\n- `404` - Ресурс не найден\n- `500` - Внутренняя ошибка сервера"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/api/dictionaries/{code}": {
      "get": {
        "summary": "Получить словарь по коду",
        "description": "Возвращает информацию о словаре.",
        "tags": [
          "Словари"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Уникальный идентификатор"
                    },
                    "code": {
                      "type": "string",
                      "description": "Код словаря"
                    },
                    "name": {
                      "type": "string",
                      "description": "Название"
                    },
                    "description": {
                      "type": "string"
                    },
                    "allowHierarchy": {
                      "type": "boolean",
                      "description": "Разрешена ли иерархия"
                    },
                    "maxDepth": {
                      "type": "number",
                      "description": "Максимальная глубина"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string",
                            "description": "Уникальный код поля",
                            "minLength": 1,
                            "pattern": "^[a-z][a-z0-9_]*$"
                          },
                          "name": {
                            "type": "string",
                            "description": "Отображаемое название",
                            "minLength": 1
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "string",
                              "number",
                              "boolean",
                              "color",
                              "icon"
                            ],
                            "description": "Тип поля"
                          },
                          "required": {
                            "type": "boolean",
                            "default": false
                          }
                        },
                        "required": [
                          "code",
                          "name",
                          "type"
                        ],
                        "description": "Дополнительное поле словаря"
                      },
                      "description": "Дополнительные поля"
                    },
                    "itemCount": {
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата создания"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата обновления"
                    }
                  },
                  "required": [
                    "id",
                    "code",
                    "name",
                    "allowHierarchy",
                    "maxDepth",
                    "fields",
                    "createdAt",
                    "updatedAt"
                  ],
                  "description": "Словарь"
                },
                "example": {
                  "id": "clx1234567890abcdef",
                  "code": "contact_types",
                  "name": "Типы контактов",
                  "description": "Словарь типов контактов для классификации",
                  "isSystem": false,
                  "isHierarchical": false,
                  "createdAt": "2024-01-05T12:00:00.000Z",
                  "updatedAt": "2024-01-05T12:00:00.000Z",
                  "_count": {
                    "items": 5
                  }
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Код словаря"
          }
        ]
      },
      "patch": {
        "summary": "Обновить словарь",
        "description": "Обновляет настройки словаря.",
        "tags": [
          "Словари"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Уникальный идентификатор"
                    },
                    "code": {
                      "type": "string",
                      "description": "Код словаря"
                    },
                    "name": {
                      "type": "string",
                      "description": "Название"
                    },
                    "description": {
                      "type": "string"
                    },
                    "allowHierarchy": {
                      "type": "boolean",
                      "description": "Разрешена ли иерархия"
                    },
                    "maxDepth": {
                      "type": "number",
                      "description": "Максимальная глубина"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string",
                            "description": "Уникальный код поля",
                            "minLength": 1,
                            "pattern": "^[a-z][a-z0-9_]*$"
                          },
                          "name": {
                            "type": "string",
                            "description": "Отображаемое название",
                            "minLength": 1
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "string",
                              "number",
                              "boolean",
                              "color",
                              "icon"
                            ],
                            "description": "Тип поля"
                          },
                          "required": {
                            "type": "boolean",
                            "default": false
                          }
                        },
                        "required": [
                          "code",
                          "name",
                          "type"
                        ],
                        "description": "Дополнительное поле словаря"
                      },
                      "description": "Дополнительные поля"
                    },
                    "itemCount": {
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата создания"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата обновления"
                    }
                  },
                  "required": [
                    "id",
                    "code",
                    "name",
                    "allowHierarchy",
                    "maxDepth",
                    "fields",
                    "createdAt",
                    "updatedAt"
                  ],
                  "description": "Словарь"
                },
                "example": {
                  "id": "clx1234567890abcdef",
                  "code": "contact_types",
                  "name": "Отрасли экономики",
                  "description": "Обновлённое описание отраслей",
                  "isSystem": false,
                  "isHierarchical": false,
                  "createdAt": "2024-01-05T12:00:00.000Z",
                  "updatedAt": "2024-01-05T12:00:00.000Z",
                  "_count": {
                    "items": 5
                  }
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Код словаря"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "allowHierarchy": {
                    "type": "boolean"
                  },
                  "maxDepth": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 3
                  },
                  "fields": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Уникальный код поля",
                          "minLength": 1,
                          "pattern": "^[a-z][a-z0-9_]*$"
                        },
                        "name": {
                          "type": "string",
                          "description": "Отображаемое название",
                          "minLength": 1
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "string",
                            "number",
                            "boolean",
                            "color",
                            "icon"
                          ],
                          "description": "Тип поля"
                        },
                        "required": {
                          "type": "boolean",
                          "default": false
                        }
                      },
                      "required": [
                        "code",
                        "name",
                        "type"
                      ],
                      "description": "Дополнительное поле словаря"
                    }
                  }
                },
                "description": "Данные для обновления словаря"
              },
              "example": {
                "name": "Отрасли экономики",
                "description": "Обновлённое описание отраслей"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Удалить словарь",
        "description": "Удаляет словарь вместе со всеми его элементами.",
        "tags": [
          "Словари"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ"
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Код словаря"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}